fdsfdsf

+++ app/utils/function.js
@@ -54,7 +54,7 @@ function VerifyRefreshToken(token) {
 
 function deleteFileInPublic(fileAddress) {
     const pathFile = path.join(__dirname, "..", "..", "public", fileAddress);
-    fs.unlinkSync(pathFile);
+    if (fs.existsSync(pathFile)) fs.unlinkSync(pathFile);
 }
 
 module.exports = {deleteFileInPublic, numberRandomGenerator, SignAccessToken, SignRefreshToken, VerifyRefreshToken};
